refactor: Calculator를 LottoService로 변경
- 역할이 도메인보다 서비스에 적합하여 변경한다.
- 내부의 메소드명을 정정한다.
