+++ src/main/java/lotto/common/config/DependencyConfig.java
@@ -1,7 +1,7 @@
 package lotto.common.config;
 
 import lotto.controller.Controller;
-import lotto.domain.Calculator;
+import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -11,8 +11,8 @@ public class DependencyConfig {
     public DependencyConfig() {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
-        Calculator calculator = new Calculator();
-        controller = new Controller(inputView, outputView, calculator);
+        LottoService lottoService = new LottoService();
+        controller = new Controller(inputView, outputView, lottoService);
     }
 
     public Controller getController() {
