+++ src/test/java/lotto/service/LottoServiceTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.service;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,9 +8,12 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import lotto.domain.Amount;
+import lotto.domain.Lotto;
+import lotto.domain.MatchStatistics;
 import lotto.dto.Profit;
 
-class CalculatorTest {
+class LottoServiceTest {
     @Test
     @DisplayName("로또 번호가 6개가 아니라면 예외를 발생시킨다.")
     void testLottoNumber_sizeException() {
@@ -20,14 +23,14 @@ class CalculatorTest {
 
     @Test
     @DisplayName("수익률 계산 테스트")
-    void testCalculateProfit() {
+    void testConvertToMapProfit() {
         HashMap<MatchStatistics, Integer> map = new HashMap<>();
         Amount amount = new Amount(10000);
 
         map.put(MatchStatistics.MATCH_THREE, 3);
 
-        Calculator calculator = new Calculator();
-        Profit profit = calculator.calculate(map, amount);
+        LottoService lottoService = new LottoService();
+        Profit profit = lottoService.calculateProfit(map, amount);
 
         assertThat(profit.rate()).isEqualTo(1.5);
     }
